requirements.txt
faster-whisper
openai-whisper